Fix discord login issues (#106)
* Bump a lot of deps

* Adjust docker build tag generation

* Use branch and pr as ref value for tag creation

* Do not optimize images

* Adjust discord provider config for auth.js

* Add more error handling to auth.ts

* Display correct error message

+++ pages/api/delete/item.ts
@@ -1,5 +1,4 @@
-import { authOptions } from '../auth/[...nextauth]'
-import { getServerSession } from 'next-auth/next'
+import { auth } from '../../../auth'
 import { canEdit } from '../../../lib/session'
 import { deleteItem } from '../../../lib/db/items'
 import { User } from '../../../types/User'
@@ -9,8 +8,8 @@ export default async function apiDeleteItem(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
-  const session = await getServerSession(req, res, authOptions)
-  if (canEdit(session)) {
+  const session = await auth(req, res)
+  if (canEdit(session) && session !== null) {
     const d = req.body
     if (d._id !== '') {
       await deleteItem(d._id, session.user as User)
